+++ src/config.ts
@@ -1,3 +1,4 @@
+import { ConfigMap } from '@pulumi/pulumi/automation';
 import {
   getBooleanInput,
   getInput,
@@ -27,14 +28,6 @@ export function makeInstallationConfig(): rt.Result<InstallationConfig> {
   });
 }
 
-const configValueRt = rt.Dictionary(
-  rt.Record({
-    value: rt.String,
-    secret: rt.Boolean.optional(),
-  }),
-  rt.String,
-);
-
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export function makeConfig() {
   return {
@@ -53,9 +46,10 @@ export function makeConfig() {
     upsert: getBooleanInput('upsert'),
     remove: getBooleanInput('remove'),
     refresh: getBooleanInput('refresh'),
-    configMap: getYAMLInput('config-map', {
-      parser: (configMap) => configValueRt.check(configMap),
-    }),
+    // TODO: Add parser back once pulumi/pulumi#12641 is fixed.
+    // @see https://github.com/pulumi/actions/pull/913
+    // @see https://github.com/pulumi/actions/pull/912
+    configMap: getYAMLInput<ConfigMap>('config-map'),
     editCommentOnPr: getBooleanInput('edit-pr-comment'),
 
     options: {